# docs/swagger.yaml
basePath: /
...
info:
  contact:
    name: Fern Platform Team
    url: https://github.com/guidewire-oss/fern-platform/issues
  description: Unified test intelligence platform — aggregate, analyze and act on
    test results from any CI/CD pipeline.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/guidewire-oss/fern-platform
  title: Fern Platform API
  version: "1.0"
paths:
...
  /api/v1/flaky-tests:
    get:
      description: Returns tests detected as intermittently failing for a project
      parameters:
      - description: Project ID to query flaky tests for
        in: query
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List flaky tests
      tags:
      - flaky-tests
  /api/v1/flaky-tests/{id}/ignore:
    post:
      description: Suppresses alerts for a detected flaky test
      parameters:
      - description: Flaky test ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ignore a flaky test
      tags:
      - flaky-tests
  /api/v1/flaky-tests/{id}/resolve:
    post:
      description: Marks a detected flaky test as resolved
      parameters:
      - description: Flaky test ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Resolve a flaky test
      tags:
      - flaky-tests
...
securityDefinitions:
  BearerAuth:
    description: 'JWT bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...